An LGPL template program in Java.
It's a simple piece of Java code (a Hello World program, to be precise)
licensed under the LGPL. The idea is that it virtually eliminates my
overhead when creating a new project using Java, Git, and the LGPL.
Forking this lets me write code that's both executable and pushable at
any time, and automatically has license text in the right places. All I
really have to do to make it official is give a descriptive name to the
repository, bottom-level folder, and primary class, and change or remove
the print statement in the primary class's main function.

I don't want to put all my code in one Git project, but starting
projects manually has proven to be time-consuming. I'm hoping the
forking process will be simpler.